+++ R/read_pheno.R
@@ -1,8 +1,8 @@
 #' Read phenotype data
 #'
 #' Read phenotype data from a CSV file (and, optionally, phenotype
-#' covariate data from a separate zip file). The CSV files may be
-#' contained in zip files, separately or togther.
+#' covariate data from a separate CSV file). The CSV files may be
+#' contained in zip files, separately or together.
 #'
 #' @md
 #'